#173: Helper Methods Refactoring
As per Cynthia
